pyproject.toml
[tool.poetry]
name = "budapest_pythonweekend"
version = "0.1.0"
description = ""
authors = ["Balazs Heizer <heizer.balazs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
pre-commit = "^2.17.0"
redis = "^4.1.4"
python-slugify = "^6.1.1"
SQLAlchemy = "^1.4.31"
psycopg2-binary = "^2.9.3"
fastapi = "^0.75.0"
uvicorn = "^0.17.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
